moved summary report download logic to services
